Properly merge Paratext metadata updates given re-arch
