+++ src/Serval/src/Serval.DataFiles/Usings.cs
@@ -26,3 +26,4 @@ global using Serval.Shared.Models;
 global using Serval.Shared.Services;
 global using Serval.Shared.Utils;
 global using SIL.DataAccess;
+global using SIL.Machine.Corpora;
